moved everythng into package to make it possible to use in another go code
